+++ app/Http/Livewire/App/Events/Tickets.php
@@ -93,6 +93,8 @@ class Tickets extends Component
 
     public function reserve()
     {
+        if($this->order !== null) return redirect()->route('app.orders.show', $this->order);
+
         $this->checkValidation();
 
         $reservation = Order::create(['event_id' => $this->event->id, 'user_id' => auth()->id(), 'reservation_ends' => now()->addDays($this->event->settings->reservation_length)]);